feat: modal adicionar item avulso
