--- app/src/main/java/com/anotasmart/ui/screens/VendaScreen.kt
@@ -1,275 +1,0 @@
-package com.anotasmart.ui.screens
-import android.annotation.SuppressLint
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Inventory2
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.anotasmart.model.entity.Category
-import com.anotasmart.model.entity.Product
-import com.anotasmart.ui.viewModels.VendaViewModel
-
-@Composable
-fun VendaScreen(
-    viewModel: VendaViewModel = viewModel()
-) {
-    val produtos by viewModel.produtos.collectAsState()
-    val categorias by viewModel.categorias.collectAsState()
-    val categoriaSelecionada by viewModel.categoriaSelecionada.collectAsState()
-    val searchQuery by viewModel.searchQuery.collectAsState()
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.background)
-    ) {
-        BarraDeBusca(
-            query = searchQuery,
-            onQueryChange = viewModel::onSearchQueryChanged
-        )
-        BotaoVendaAvulsa(
-            onClick = { }
-        )
-        ListaCategorias(
-            categorias = categorias,
-            categoriaSelecionadaId = categoriaSelecionada,
-            onCategoriaClick = viewModel::onCategoriaSelecionada
-        )
-        GradeProdutos(
-            produtos = produtos,
-            onProdutoClick = { produto -> {} }
-        )
-    }
-}
-
-
-@Composable
-fun BarraDeBusca(query: String, onQueryChange: (String) -> Unit) {
-    OutlinedTextField(
-        value = query,
-        onValueChange = onQueryChange,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 8.dp),
-        placeholder = { Text("Buscar") },
-        leadingIcon = {
-            Icon(
-                imageVector = Icons.Default.Search,
-                contentDescription = "Ícone de busca"
-            )
-        },
-        singleLine = true,
-        shape = RoundedCornerShape(8.dp)
-    )
-}
-
-@Composable
-fun BotaoVendaAvulsa(onClick: () -> Unit) {
-    Button(
-        onClick = onClick,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 8.dp)
-            .height(50.dp),
-        shape = RoundedCornerShape(8.dp),
-        colors = ButtonDefaults.buttonColors(
-            containerColor = MaterialTheme.colorScheme.secondaryContainer,
-            contentColor = MaterialTheme.colorScheme.onSecondaryContainer
-        )
-    ) {
-        Icon(
-            imageVector = Icons.Default.Add,
-            contentDescription = "Item avulso",
-            modifier = Modifier.padding(end = 8.dp)
-        )
-        Text(text = "ITEM AVULSO", fontWeight = FontWeight.Bold)
-    }
-}
-
-@Composable
-fun ListaCategorias(categorias: List<Category>, categoriaSelecionadaId: String, onCategoriaClick: (String) -> Unit) {
-    Box(modifier = Modifier.fillMaxWidth()) {
-        LazyRow(
-            modifier = Modifier.fillMaxWidth(),
-            contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            items(categorias) { categoria ->
-                val isSelected = categoria.id == categoriaSelecionadaId
-
-                Surface(
-                    shape = RoundedCornerShape(4.dp),
-                    color = if (isSelected) MaterialTheme.colorScheme.primary
-                            else MaterialTheme.colorScheme.surfaceVariant,
-                    contentColor = if (isSelected) MaterialTheme.colorScheme.onPrimary
-                                   else MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.clickable { onCategoriaClick(categoria.id) }
-                ) {
-                    Text(
-                        text = categoria.nome,
-                        modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-                        style = MaterialTheme.typography.labelMedium,
-                        fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Normal
-                    )
-                }
-            }
-        }
-        Box( // máscara de transparência
-            modifier = Modifier
-                .matchParentSize()
-                .background(
-                    brush = Brush.horizontalGradient(
-                        0.85f to Color.Transparent,
-                        1.0f to MaterialTheme.colorScheme.background
-                    )
-                )
-        )
-    }
-}
-
-@SuppressLint("DefaultLocale")
-@Composable
-fun ItemProduto(produto: Product, onClick: () -> Unit) {
-    Card(
-        modifier = Modifier.fillMaxWidth().clickable { onClick() },
-        shape = RoundedCornerShape(4.dp),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant)
-    ) {
-        Column(
-            modifier = Modifier.padding(8.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(100.dp)
-                    .background(MaterialTheme.colorScheme.surfaceVariant, RoundedCornerShape(4.dp)),
-                contentAlignment = Alignment.Center
-            ) {
-                if (!produto.imagePath.isNullOrEmpty()) {
-                    val imageResourceId = produto.imagePath.toIntOrNull()
-                    if (imageResourceId != null) {
-                        Image(
-                            painter = painterResource(id = imageResourceId),
-                            contentDescription = "Imagem do produto ${produto.nome}",
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(100.dp)
-                                .clip(RoundedCornerShape(4.dp)),
-                            contentScale = ContentScale.Crop
-                        )
-                    }
-                } else {
-                    // Exibe o ícone genérico caso o produto não possua imagem
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(100.dp)
-                            .background(MaterialTheme.colorScheme.surfaceVariant, RoundedCornerShape(4.dp)),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Inventory2,
-                            contentDescription = "Ícone de produto sem imagem",
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                            modifier = Modifier.size(32.dp)
-                        )
-                    }
-                }
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Text(
-                text = produto.nome,
-                style = MaterialTheme.typography.labelSmall,
-                fontWeight = FontWeight.Bold,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                textAlign = TextAlign.Center,
-            )
-
-            Text(
-                text = "R$ ${String.format("%.2f", produto.precoVenda)}",
-                style = MaterialTheme.typography.bodyMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.primary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-
-            Text(
-                text = "${produto.quantidadeEstoque} ${produto.unidadeMedida.name}",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-        }
-    }
-}
-
-@Composable
-fun GradeProdutos(
-    produtos: List<Product>,
-    onProdutoClick: (Product) -> Unit
-) {
-    LazyVerticalGrid(
-        columns = GridCells.Fixed(3),
-        contentPadding = PaddingValues(16.dp),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-        modifier = Modifier.fillMaxSize()
-    ) {
-        items(produtos) { produto ->
-            ItemProduto(
-                produto = produto,
-                onClick = { onProdutoClick(produto) }
-            )
-        }
-    }
-}
