+++ app/src/main/java/com/anotasmart/ui/screens/documentacao/DocumentacaoScreen.kt
@@ -1,4 +1,4 @@
-package com.anotasmart.ui.screens
+package com.anotasmart.ui.screens.documentacao
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,8 +8,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun MaisScreen() {
+fun DocumentacaoScreen() {
     Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text(text = "Tela de Mais")
+        Text(text = "Tela de Documentação")
     }
 }
