+++ app/src/main/java/com/anotasmart/ui/screens/vendas/components/DialogItemAvulso.kt
@@ -1,0 +1,131 @@
+package com.anotasmart.ui.screens.vendas.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Remove
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableDoubleStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun DialogAdicionarCarrinhoItemAvulso(
+    onDismissRequest: () -> Unit,
+    onConfirmar: (precoCusto: Double?, precoVenda: Double, quantidade: Double) -> Unit
+) {
+    var precoCusto by remember { mutableStateOf("") }
+    var precoVenda by remember { mutableStateOf("") }
+    var quantidadeSelecionada by remember { mutableDoubleStateOf(1.0) }
+
+    val pVenda = precoVenda.toDoubleOrNull() ?: 0.0
+    val valorTotal = pVenda * quantidadeSelecionada
+
+    AlertDialog(
+        onDismissRequest = onDismissRequest,
+        title = {
+            Text(
+                text = "Item Avulso",
+                fontWeight = FontWeight.Bold,
+                textAlign = TextAlign.Center,
+                modifier = Modifier.fillMaxWidth()
+            )
+        },
+        text = {
+            Column(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                OutlinedTextField(
+                    value = precoVenda,
+                    onValueChange = { precoVenda = it },
+                    label = { Text("Preço de Venda") },
+                    modifier = Modifier.fillMaxWidth(),
+                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Decimal),
+                    prefix = { Text("R$ ") }
+                )
+
+                Spacer(modifier = Modifier.height(12.dp))
+
+                OutlinedTextField(
+                    value = precoCusto,
+                    onValueChange = { precoCusto = it },
+                    label = { Text("Preço de Custo (Opcional)") },
+                    modifier = Modifier.fillMaxWidth(),
+                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Decimal),
+                    prefix = { Text("R$ ") }
+                )
+
+                Spacer(modifier = Modifier.height(24.dp))
+
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.Center,
+                    modifier = Modifier.fillMaxWidth()
+                ) {
+                    IconButton(
+                        onClick = { if (quantidadeSelecionada > 1.0) quantidadeSelecionada -= 1.0 },
+                        enabled = quantidadeSelecionada > 1.0
+                    ) {
+                        Icon(imageVector = Icons.Default.Remove, contentDescription = "Diminuir quantidade")
+                    }
+
+                    Text(
+                        text = quantidadeSelecionada.toString(),
+                        style = MaterialTheme.typography.headlineMedium,
+                        modifier = Modifier.padding(horizontal = 24.dp)
+                    )
+
+                    IconButton(
+                        onClick = { quantidadeSelecionada += 1.0 }
+                    ) {
+                        Icon(imageVector = Icons.Default.Add, contentDescription = "Aumentar quantidade")
+                    }
+                }
+            }
+        },
+        confirmButton = {
+            Box(
+                modifier = Modifier.fillMaxWidth(),
+                contentAlignment = Alignment.Center
+            ) {
+                Button(
+                    onClick = {
+                        val pCusto = precoCusto.toDoubleOrNull()
+                        onConfirmar(pCusto, pVenda, quantidadeSelecionada)
+                    },
+                    modifier = Modifier.fillMaxWidth(),
+                    enabled = precoVenda.isNotEmpty() && pVenda > 0
+                ) {
+                    Text(
+                        text = "ADICIONAR R$ ${String.format("%.2f", valorTotal)}",
+                        fontWeight = FontWeight.Bold
+                    )
+                }
+            }
+        }
+    )
+}
