+++ app/src/main/java/com/anotasmart/ui/screens/vendas/components/GradeProdutos.kt
@@ -1,0 +1,33 @@
+package com.anotasmart.ui.screens.vendas.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.anotasmart.model.entity.Product
+
+@Composable
+fun GradeProdutos(
+    produtos: List<Product>,
+    onProdutoClick: (Product) -> Unit
+) {
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(3),
+        contentPadding = PaddingValues(16.dp),
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+        modifier = Modifier.fillMaxSize()
+    ) {
+        items(produtos) { produto ->
+            ItemProduto(
+                produto = produto,
+                onClick = { onProdutoClick(produto) }
+            )
+        }
+    }
+}
