+++ app/src/main/java/com/anotasmart/ui/viewModels/VendaViewModel.kt
@@ -1,7 +1,7 @@
 package com.anotasmart.ui.viewModels
 
 import androidx.lifecycle.ViewModel
-import com.anotasmart.ui.screens.MockDataSource
+import com.anotasmart.data.mocks.MockDataSource
 import com.anotasmart.model.entity.Category
 import com.anotasmart.model.entity.Product
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -11,15 +11,16 @@ import kotlinx.coroutines.flow.asStateFlow
 class VendaViewModel : ViewModel() {
     private val _produtos = MutableStateFlow<List<Product>>(emptyList())
     val produtos: StateFlow<List<Product>> = _produtos.asStateFlow()
-
     private val _categorias = MutableStateFlow<List<Category>>(emptyList())
     val categorias: StateFlow<List<Category>> = _categorias.asStateFlow()
-
     private val _categoriaSelecionada = MutableStateFlow("1")
     val categoriaSelecionada: StateFlow<String> = _categoriaSelecionada.asStateFlow()
-
     private val _searchQuery = MutableStateFlow("")
     val searchQuery: StateFlow<String> = _searchQuery.asStateFlow()
+    private val _produtoSelecionado = MutableStateFlow<Product?>(null)
+    val produtoSelecionado: StateFlow<Product?> = _produtoSelecionado.asStateFlow()
+    private val _mostrarDialogItemAvulso = MutableStateFlow(false)
+    val mostrarDialogItemAvulso: StateFlow<Boolean> = _mostrarDialogItemAvulso.asStateFlow()
 
     init {
         carregarDadosMock()
@@ -36,5 +37,31 @@ class VendaViewModel : ViewModel() {
 
     fun onCategoriaSelecionada(id: String) {
         _categoriaSelecionada.value = id
+    }
+
+    fun selecionarProdutoParaCarrinho(produto: Product) {
+        _produtoSelecionado.value = produto
+    }
+
+    fun limparProdutoSelecionado() {
+        _produtoSelecionado.value = null
+    }
+
+    fun abrirDialogItemAvulso() {
+        _mostrarDialogItemAvulso.value = true
+    }
+
+    fun fecharDialogItemAvulso() {
+        _mostrarDialogItemAvulso.value = false
+    }
+
+    fun adicionarAoCarrinho(produto: Product, quantidade: Double) {
+        // lógica
+        limparProdutoSelecionado()
+    }
+
+    fun adicionarItemAvulsoAoCarrinho(precoCusto: Double?, precoVenda: Double, quantidade: Double) {
+        // lógica para adicionar item avulso
+        fecharDialogItemAvulso()
     }
 }
